// language-learning-app-frontend-main/src/api/model/wordResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * German Language Learning API
 * API for managing German words, phrases, components, examples, and semantic groups.
 * OpenAPI spec version: 1.0.0
 */
import type { WordResponseReflexivity } from './wordResponseReflexivity';
import type { WordResponseCase } from './wordResponseCase';
import type { WordResponsePartOfSpeech } from './wordResponsePartOfSpeech';
import type { WordResponsePartOfSpeechObj } from './wordResponsePartOfSpeechObj';
import type { WordResponseCategories } from './wordResponseCategories';
import type { WordResponseCategoryNames } from './wordResponseCategoryNames';
import type { WordResponseLevel } from './wordResponseLevel';
import type { WordResponseFrequency } from './wordResponseFrequency';
import type { WordComponentResponse } from './wordComponentResponse';
import type { WordMeaningResponse } from './wordMeaningResponse';
import type { WordResponseGender } from './wordResponseGender';
import type { WordResponsePluralForm } from './wordResponsePluralForm';
import type { WordResponseVerbForm2 } from './wordResponseVerbForm2';
import type { WordResponseVerbForm3 } from './wordResponseVerbForm3';

export interface WordResponse {
  id: number;
  text: string;
  reflexivity: WordResponseReflexivity;
  case: WordResponseCase;
  part_of_speech?: WordResponsePartOfSpeech;
  part_of_speech_obj?: WordResponsePartOfSpeechObj;
  categories?: WordResponseCategories;
  category_names?: WordResponseCategoryNames;
  level?: WordResponseLevel;
  frequency?: WordResponseFrequency;
  components: WordComponentResponse[];
  component_count: number;
  value_count: number;
  meanings?: WordMeaningResponse[];
  gender?: WordResponseGender;
  plural_form?: WordResponsePluralForm;
  verb_form2?: WordResponseVerbForm2;
  verb_form3?: WordResponseVerbForm3;
}